# gomino-src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  media.UploadResponse:
    properties:
      mediaValue:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      api:message:
        type: string
      api:statuscode:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gomino.com
    name: API Support
  description: This is the API server for GoMino.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoMino API
  version: "1.0"
paths:
...
  /g/s/media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload media (image/video) and convert to AVIF
      parameters:
      - description: Media file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/media.UploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Upload media file
      tags:
      - media
...
swagger: "2.0"